refactor: nomenclature & classifier model refactor: task scheme refactor:  nomenclature scheme
