+++ src/scheme/embedding/embedding_scheme.py
@@ -1,0 +1,16 @@
+from rq.job import JobStatus
+from sqlmodel import SQLModel
+
+
+class CreateAndSaveEmbeddingsUpload(SQLModel):
+    db_con_str: str
+    table_name: str
+    collection_name: str
+    chunk_size: int | None = 500
+
+
+class CreateAndSaveEmbeddingsResult(SQLModel):
+    job_id: str
+    status: JobStatus
+    ready_count: int | None
+    total_count: int | None
